* '4.7' of scm.dev.nokia.troll.no:qt/oslo-staging-2: (62 commits)
  Fix warnings
  Fix shadow building of spectrum demo
  Doc: Removed a link to missing external online documentation.
  Fixes QGraphicsItem::scroll issues
  Fixes wrong QGraphicsItemCache::boundingRect.
  changes-4.7.0 updated
  Revert "Deselect the current selection when the QItemSelectionModel::model is reset."
  add missing include
  Fix crash when using fonts in non-gui QApplication
  Deselect the current selection when the QItemSelectionModel::model is reset.
  Autotest fix on macosx
  Compile with gcc 4.0.1
  improve Unicode Normalization autotest
  more subtests for QChar
  nano optimization of canonicalOrderHelper()
  fix canonicalOrderHelper() for some corner case
  use new QChar::requiresSurrogates() instead of hardcoded value
  prevent fake normalization
  prefer QChar::*surrogate() over hardcoded values
  Doc: Updated example license to three clause BSD license.
  ...

+++ src/corelib/thread/qreadwritelock.cpp
@@ -252,7 +252,7 @@ bool QReadWriteLock::tryLockForRead(int timeout)
 
     while (d->accessCount < 0 || d->waitingWriters) {
         ++d->waitingReaders;
-        bool success = d->readerWait.wait(&d->mutex, timeout < 0 ? ULONG_MAX : timeout);
+        bool success = d->readerWait.wait(&d->mutex, timeout < 0 ? ULONG_MAX : ulong(timeout));
         --d->waitingReaders;
         if (!success)
             return false;
@@ -374,7 +374,7 @@ bool QReadWriteLock::tryLockForWrite(int timeout)
 
     while (d->accessCount != 0) {
         ++d->waitingWriters;
-        bool success = d->writerWait.wait(&d->mutex, timeout < 0 ? ULONG_MAX : timeout);
+        bool success = d->writerWait.wait(&d->mutex, timeout < 0 ? ULONG_MAX : ulong(timeout));
         --d->waitingWriters;
 
         if (!success)
